pyproject.toml
[tool.poetry]
name = "notify-bot"
version = "0.1.0"
description = ""
authors = ["Konstantin Raikhert <raikhert13@gmail.com>"]
readme = "README.md"
packages = [{include = "notify_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
APScheduler = "^3.10.1"
aiogram = "3.0.0b7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
